+++ EmpiriaBMS.Models/Models/Client.cs
@@ -15,7 +15,7 @@ public class Client : Entity
 
     public string? Description { get; set; }
 
-    public int? AddressId { get; set; }
+    public long? AddressId { get; set; }
     public Address? Address { get; set; }
 
     public double PotencialFee { get; set; }
