+++ EmpiriaBMS.Models/Models/Deliverable.cs
@@ -6,11 +6,11 @@ namespace EmpiriaBMS.Models.Models;
 public class Deliverable : Entity
 {
     [Required]
-    public int TypeId { get; set; }
+    public long TypeId { get; set; }
     public DeliverableType? Type { get; set; }
 
     [Required]
-    public int DisciplineId { get; set; }
+    public long DisciplineId { get; set; }
     public Discipline? Discipline { get; set; }
 
     public float CompletionEstimation { get; set; }
