+++ EmpiriaBMS.Models/Models/UserRole.cs
@@ -2,9 +2,9 @@
 namespace EmpiriaBMS.Models.Models;
 public class UserRole : Entity
 {
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public User? User { get; set; }
 
-    public int RoleId { get; set; }
+    public long RoleId { get; set; }
     public Role? Role { get; set; }
 }
